requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.49.0
dvc-data==3.15.1
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
filelock==3.13.4
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grandalf==0.8
gto==1.7.1
httplib2==0.22.0
hydra-core==1.3.2
idna==3.7
imbalanced-learn==0.12.2
imblearn==0.0
iterative-telemetry==0.0.8
joblib==1.4.0
kiwisolver==1.4.5
kombu==5.3.7
markdown-it-py==3.0.0
matplotlib==3.8.4
mdurl==0.1.2
multidict==6.0.5
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.10.0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
scmrepo==3.3.1
seaborn==0.13.2
semver==3.0.2
setuptools==69.5.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
sklearn-features==0.0.2
smmap==5.0.1
sqltrie==0.11.0
tabulate==0.9.0
threadpoolctl==3.4.0
tomlkit==0.12.4
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
yarl==1.9.4
zc.lockfile==3.0.post1
